LayoutTests/fast/dom/Window/extrazoom/window-get-real-screen-dimensions-expected.txt

PASS screen.width is 320
PASS screen.height is 568
PASS successfullyParsed is true

TEST COMPLETE
